+++ todo-app/bin/Debug/net9.0/todo-app.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v9.0",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v9.0": {
+      "todo-app/1.0.0": {
+        "runtime": {
+          "todo-app.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "todo-app/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
